requirements.txt
streamlit 
langchain 
langchain_groq
selenium
beautifulsoup4
python-dotenv